requirements.txt
llama-cpp-python==0.3.10
langchain==0.1.14
langchain-community
langchain-core
faiss-cpu
sentence-transformers
torch
transformers
numpy
jinja2
diskcache